improved ux
creating custom parts now creates a part called Custom Part (number here). It automatically brings the user to the part editor screen. The images now act as buttons to take you to the part editor in the add parts screen like the images in the normal parts screen. Hovering makes it slightly grayish. It now uses a class called custom parts for the styling.